% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBlockDataframe.R,
%   R/CreateBlockDataframe_MotorTasks.R
\name{CreateBlockDataframe_MotorTasks}
\alias{CreateBlockDataframe_MotorTasks}
\title{Extract block data from json file}
\usage{
CreateBlockDataframe_MotorTasks(d)

CreateBlockDataframe_MotorTasks(d)
}
\arguments{
\item{json}{A json file imported via jsonlite::read_json()}
}
\value{
A dataframe with task data extracted from json

A dataframe with task data extracted from json
}
\description{
Internal function for extracting block data from json files exported from
MRL1/MRL2/VMR/MRV tasks.

Internal function for extracting block data from json files exported from
MRL1/MRL2/VMR/MRV tasks.
}
