% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSubjectDataframe_PALTask.R
\name{CreateSubjectDataframe_PALTask}
\alias{CreateSubjectDataframe_PALTask}
\title{Extract subject data from json file}
\usage{
CreateSubjectDataframe_PALTask(json)
}
\arguments{
\item{json}{A json file imported via jsonlite::read_json()}
}
\value{
A dataframe with subject data extracted from json
}
\description{
Internal function for extracting block data from json files exported from
PAL task.
}
