% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskInIGT.R
\name{ImportTaskInIGT}
\alias{ImportTaskInIGT}
\title{Extract IGT data from Inquisit file}
\usage{
ImportTaskInIGT(file)
}
\arguments{
\item{json}{A path to a inquisit file}
}
\value{
A dataframe
}
\description{
Function for extracting data from inquisit files exported from the IGT.
}
