% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskInST.R
\name{ImportTaskInST}
\alias{ImportTaskInST}
\title{Extract Simon data from Inquisit file}
\usage{
ImportTaskInST(file)
}
\arguments{
\item{json}{A path to a inquisit file}
}
\value{
A dataframe
}
\description{
Function for extracting data from inquisit files exported from the Simon task.
}
