% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskInTMT.R
\name{ImportTaskInTMT}
\alias{ImportTaskInTMT}
\title{Extract Trail Making data from Inquisit file}
\usage{
ImportTaskInTMT(file)
}
\arguments{
\item{json}{A path to a inquisit file}
}
\value{
A dataframe
}
\description{
Function for extracting data from inquisit files exported from the Train Making task.
}
