% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskInWP.R
\name{ImportTaskInWP}
\alias{ImportTaskInWP}
\title{Extract weather prediction data from Inquisit file}
\usage{
ImportTaskInWP(file)
}
\arguments{
\item{json}{A path to a inquisit file}
}
\value{
A dataframe
}
\description{
Function for extracting data from inquisit files exported from the
weather prediction task.
}
