% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskMRL1.R
\name{ImportTaskMRL1}
\alias{ImportTaskMRL1}
\title{Extract MRL1 data from json file}
\usage{
ImportTaskMRL1(json)
}
\arguments{
\item{json}{A path to a .json file}
}
\value{
A list of four data frames
}
\description{
Function for extracting data from json files exported from the MRL1 task.
Parses the json file and returns a list containing subject data, system data,
global task variables, and behavior.
}
