% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskPAL.R
\name{ImportTaskPAL}
\alias{ImportTaskPAL}
\title{Extract PAL data from json file}
\usage{
ImportTaskPAL(json)
}
\arguments{
\item{json}{A path to a .json file}
}
\value{
A list of four data frames
}
\description{
Function for extracting data from json files exported from the PAL task.
Parses the json file and returns a list containing global task variables
and behavior.
}
