% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskRL1.R
\name{ImportTaskRL1}
\alias{ImportTaskRL1}
\title{Extract RL1 data from json file}
\usage{
ImportTaskRL1(json)
}
\arguments{
\item{json}{A path to a .json file}
}
\value{
A list of four data frames
}
\description{
Function for extracting data from json files exported from the RL1 task.
Parses the json file and returns a list containing subject data, system data,
global task variables, and behavior.
}
