% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskMRL2.R
\name{ImportTaskRL2}
\alias{ImportTaskRL2}
\title{Extract RL2 data from json file}
\usage{
ImportTaskRL2(json)
}
\arguments{
\item{json}{A path to a .json file}
}
\value{
A list of four data frames
}
\description{
Function for extracting data from json files exported from the RL2 task.
Parses the json file and returns a list containing subject data, system data,
global task variables, and behavior.
}
