% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportTaskVMR.R
\name{ImportTaskVMR}
\alias{ImportTaskVMR}
\title{Extract VMR data from json file}
\usage{
ImportTaskVMR(json)
}
\arguments{
\item{json}{A path to a .json file}
}
\value{
A list of four data frames
}
\description{
Function for extracting data from json files exported from the VMR task.
Parses the json file and returns a list containing subject data, system data,
global task variables, and behavior.
}
