% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizeTaskMRL1.R
\name{OrganizeTaskMRL1}
\alias{OrganizeTaskMRL1}
\title{Organize MRL1 data after import}
\usage{
OrganizeTaskMRL1(data)
}
\arguments{
\item{json}{A list returned by ImportTaskMRL1}
}
\value{
A dataframe
}
\description{
Function accepts a list returned by ImportTaskMRL1 and returns a
dataframe of subject task data
}
