% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizeTaskMRL2.R
\name{OrganizeTaskMRL2}
\alias{OrganizeTaskMRL2}
\title{Organize MRL2 data after import}
\usage{
OrganizeTaskMRL2(data)
}
\arguments{
\item{json}{A list returned by ImportTaskMRL2}
}
\value{
A dataframe
}
\description{
Function accepts a list returned by ImportTaskMRL2 and returns a
dataframe of subject task data
}
