% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizeTaskRL1.R
\name{OrganizeTaskRL1}
\alias{OrganizeTaskRL1}
\title{Organize RL1 data after import}
\usage{
OrganizeTaskRL1(data)
}
\arguments{
\item{json}{A list returned by ImportTaskVMR}
}
\value{
A dataframe
}
\description{
Function accepts a list returned by ImportTaskRL1 and returns a
dataframe of subject task data
}
