% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizeTaskVMR.R
\name{OrganizeTaskVMR}
\alias{OrganizeTaskVMR}
\title{Organize VMR data after import}
\usage{
OrganizeTaskVMR(data)
}
\arguments{
\item{json}{A list returned by ImportTaskVMR}
}
\value{
A dataframe
}
\description{
Function accepts a list returned by ImportTaskVMR and returns a
dataframe of subject task data
}
