% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTaskIGT.R
\name{PlotTaskIGT}
\alias{PlotTaskIGT}
\title{Plot subject IGT data}
\usage{
PlotTaskIGT(data, plot = F)
}
\arguments{
\item{data}{A dataframe returned by ImportTaskInIGT}

\item{plot}{Logical. If true, return a ggplot object. If false, returns a dataframe.}
}
\value{
Either a ggplot object or a dataframe
}
\description{
Function accepts a a dataframe output by ImportTaskInIGT and plots choice
and reaction time.
}
