% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTaskMRL2.R
\name{PlotTaskMRL2}
\alias{PlotTaskMRL2}
\title{Plot subject MRL2 data}
\usage{
PlotTaskMRL2(data, plot = T)
}
\arguments{
\item{data}{A dataframe returned by OrganizeTaskMRL1}

\item{plot}{Logical. If true, return a ggplot object. If false, returns a dataframe.}
}
\value{
Either a ggplot object or a dataframe
}
\description{
Function accepts a a dataframe output by OrganizeDataMRL1 and plots score
and reaction time.
}
