% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTaskRL1.R
\name{PlotTaskRL1}
\alias{PlotTaskRL1}
\title{Plot subject RL1 data}
\usage{
PlotTaskRL1(data, plot = F)
}
\arguments{
\item{data}{A dataframe returned by OrganizeTaskRL1}

\item{plot}{Logical. If true, return a ggplot object. If false, returns a dataframe.}
}
\value{
Either a ggplot object or a dataframe
}
\description{
Function accepts a a dataframe output by OrganizeDataRL1 and plots score
and reaction time.
}
