% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTaskST.R
\name{PlotTaskST}
\alias{PlotTaskST}
\title{Plot subject ST data}
\usage{
PlotTaskST(data, plot = F)
}
\arguments{
\item{data}{A dataframe returned by ImportTaskInST}

\item{plot}{Logical. If true, return a ggplot object. If false, returns a dataframe.}
}
\value{
Either a ggplot object or a dataframe
}
\description{
Function accepts a a dataframe output by ImportTaskInST and plots choice
and reaction time.
}
