% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTaskVMR.R
\name{PlotTaskVMR}
\alias{PlotTaskVMR}
\title{Plot subject VMR data}
\usage{
PlotTaskVMR(data, plot = F)
}
\arguments{
\item{data}{A dataframe returned by OrganizeTaskVMR}

\item{plot}{Logical. If true, return a ggplot object. If false, returns a dataframe.}
}
\value{
Either a ggplot object or a dataframe
}
\description{
Function accepts a a dataframe output by OrganizeDataVMR and plots error
and reaction time.
}
